What is Swords and Axes
Sword's and Axe's is a 2D melee fighting game similar to Tekken, and
Street fighter.

Featuring 3 character's with the possibility of a 4th our characters
will use their traditional weapons to hack at each other till only one
stands.

This simple 2D game is very different from Tekken and Street fighter due
to the fact that a single well placed hit can kill an your adversary
while a well played adversary can out last any attack if defending
properly.

The aim of this project is to make a game that can be played on
computers and in the future even tablets that use android.